Add start of spawner and enemy logic
Towers now follow enemies based on the events recieved from the spawner.
Eventually I will have to add a range around the tower, so that it can
only follow enemies that are close to it.

This event based idea might be done in the opposite way in the future:
Towers asks the spawner if there is an enemy in its vicinity, spawner
replies back with an enemy or not (might not be the best solution, but
it's an idea).