Implement simple shooting logic
Doesn't deal damage yet and the turrets can only aim at one single thing
at a time.

Ideally I want to improve this by first looking which enemy is closest.
It would also be nice to have a pool of all gameobjects where the class
can check for hit detections (it is now done in the tower object).