Add 'Mega' variants of the towers
When placing the same towers in a 2x2 grid, there is an option with the
'q' key to upgrade them to a mega tower. For now its damage is
quadrupled and range is doubled and firerate is also quadrupled.

In the future this tower will be capable of doing more than this.