Add new map asset
This file is now in a JSON format, that I (unfortunately) edited by
hand. This should be automated in the future, with an included map
editor.

The JSON is laid out as follows: An array of arrays, these arrays define
the rows. Inside each row are the columns. Inside each JSON objects are
the different layers where we can specify an x and y which correspond to
the sprite index in the spritesheet.