Fix turret targeting bug
Turned out a simple fix after all, though very annoying and hard to
find this one. This was caused by commit 8f0d62f01008fe77d54253bb6ebe2977ff574b26.

In this commit I added a `SIZE` constant, but I also replaced the `size` property
with nothing. Later in `gameobject.lua` the size is set to a default size (something
with scale).

All that was needed to do to fix this was to add this property back.