Fix `:isSearchable` implementation
Not cleanest solution, as it is not efficient. This check runs
the search method underneath so we are probably wasting some cpu
time everytime the path gets updated. But I am okay with this for
now.

There is probably a better solution, like checking from the perspective
of the base to see if it is blocked from any direction (but that would
also require some type of a-star algorithm...).